C/Python APIs: fix memory layout for boolean
In Ada, boolean take one single byte of memory. Adapt C and Python
bindings to comply with this. This is necessary to keep structure
layouts in all languages consistent.

TN: P307-027

+++ langkit/python_api.py
@@ -87,7 +87,7 @@ class PythonAPISettings(AbstractAPISettings):
             return "_{}".format(name)
 
         return dispatch_on_type(type, [
-            (ct.BoolType, lambda _: ctype_type('c_int')),
+            (ct.BoolType, lambda _: ctype_type('c_uint8')),
             (ct.LongType, lambda _: ctype_type('c_long')),
             (ct.SourceLocationRangeType, lambda _: wrapped_type('SlocRange')),
             (ct.Token, lambda _: wrapped_type('token')),
